Fix: 최근 검색어 저장 기능 추가 (로컬스토리지 저장 hooks 및 최근 검색어 저장 hooks 생성)
